% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualnetrecurringrevenueretention.R
\name{AnnualNetRecurringRevenueRetention}
\alias{AnnualNetRecurringRevenueRetention}
\title{\code{AnnualNetRecurringRevenueRetention}}
\usage{
AnnualNetRecurringRevenueRetention(data, by = "year", ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time period used in defining the cohorts: "year", "quarter", "month", "week", day".}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A matrix
}
\description{
Computes recurring revenue, by cohort.
}
