% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bass.R
\name{Bass}
\alias{Bass}
\title{\code{Bass}}
\usage{
Bass(data, remove.last = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{remove.last}{Remove the final period (as usually is incomplete).}
}
\value{
A \code{\link{list}} containing the following elements:
\item{p}{The coefficient of innovation}.
\item{q}{The coefficient of immitation}.
\item{observed}{The observed sales/penetration}.
\item{predicted}{The predicted/fitted sales/penetration}.
\item{max}{The maximum revenue (or the revenue at 1,000 time periods, whichever is smaller)}.
}
\description{
Fits the Bass model for new products.
}
\details{
Fits the model using OLS. While NLS estimation may be better in some situations, OLS is less prone to computational errors (e.g., problems with start points, non-convergence)
}
\references{
Bass, Frank (1969). "A new product growth for model consumer durables".
Management Science 15 (5): 215-227.
}
