% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churn.R
\name{Churn}
\alias{Churn}
\title{\code{Churn}}
\usage{
Churn(data, volume = FALSE, by = "quarter", error.if.no.data = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{volume}{Weights the results by volume.}

\item{by}{The time unit to plot. E.g., "month".}

\item{error.if.no.data}{If TRUE and the data contains no valid cases, an error is thrown.}
}
\value{
A named vector showing churn.
}
\description{
Computes retention, by cohort.
}
\details{
Where subscribers suspends their purchasing for a period, 
but purchases again later, the subscriber
is included in the churn. Churn is show for all periods that have 
data in \code{data$to.period}, even if they occur in the future.
}
