% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecteddotplot.R
\name{ConnectedDotPlot}
\alias{ConnectedDotPlot}
\title{\code{ConnectedDotPlot}}
\usage{
ConnectedDotPlot(name, from, to)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}
}
\description{
Plots the subscription periods as a dotplot, with lines showing when subscriptions occurred
}

