% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetime.R
\name{CumulativeValuePlot}
\alias{CumulativeValuePlot}
\title{CumulativeValuePlot}
\usage{
CumulativeValuePlot(x)
}
\arguments{
\item{x}{A \link{\code{LifetimeValue}} object.}
}
\description{
Plots the cumulative value over time.
}

