% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churn.R
\name{CustomerChurn}
\alias{CustomerChurn}
\title{\code{CustomerChurn}}
\usage{
CustomerChurn(data, by = "quarter", ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{...}{Other parameters}
}
\value{
A named vector showing churn.
}
\description{
\code{CustomerChurn}
}
