% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churnbycohort.R
\name{CustomerChurnByCohort}
\alias{CustomerChurnByCohort}
\title{\code{CustomerChurnByCohort}}
\usage{
CustomerChurnByCohort(data, remove.last = FALSE, by, ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{remove.last}{Remove the final period (as usually is incomplete).}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A \code{\link{matrix}}
}
\description{
Computes the customer churn by time period and time-based cohort
}
