% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{CustomerGrowth}
\alias{CustomerGrowth}
\title{\code{RecurringRevenueGrowth}}
\usage{
CustomerGrowth(data, by, ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A named vector showing churn.
}
\description{
Computes retention, by cohort.
}
\details{
Where subscribers suspends their purchasing for a period, 
but purchases again later, the subscriber
is included in the churn. Churn is show for all periods that have 
data in \code{data$to.period}, even if they occur in the future.
}
