% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{Customers}
\alias{Customers}
\title{\code{Customers}}
\usage{
Customers(data, volume = FALSE, by = "quarter", error.if.no.data = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time unit to plot. E.g., "month".}
}
\value{
A named vector showing the number of customers.
}
\description{
The number of customers.
}
