% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Diagonal}
\alias{Diagonal}
\title{\code{Diagonal}}
\usage{
Diagonal(x, off = FALSE)
}
\arguments{
\item{x}{A \code{matrix} or a number indicating the number of rows in a square matrix.}

\item{off}{\code{logical}. Operates on the off diagonal if selected.}
}
\description{
Extract or replace the diagonal of a matrix, or construct a diagonal matri.
}

