% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FillInDateRowsInMatrix}
\alias{FillInDateRowsInMatrix}
\title{\code{FillInDateRowsInMatrix}}
\usage{
FillInDateRowsInMatrix(x, by, value = 0)
}
\arguments{
\item{x}{The matrix.}

\item{by}{The aggregation of the dates (e.g., "month", "year")}

\item{value}{The value to fill in for cells that are not in the origial matrix.}
}
\description{
Fills in missing date rows in a matrix.
}
