% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FillInDateVector}
\alias{FillInDateVector}
\title{\code{FillInDateVector}}
\usage{
FillInDateVector(x, by, value = 0)
}
\arguments{
\item{x}{The vector}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{value}{The value to fill in for cells that are not in the origial matrix.}
}
\description{
Fills in missing date rows in a matrix.
}
