% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FillInMatrix}
\alias{FillInMatrix}
\title{\code{FillInMatrix}}
\usage{
FillInMatrix(x, row.names, col.names, value = 0)
}
\arguments{
\item{x}{The matrix.}

\item{row.names}{The required row names, which includes the current row names
as a subset.}

\item{col.names}{The required colum names, which includes the current column names
as a subset.}

\item{value}{The value to fill in for cells that are not in the origial matrix.}
}
\description{
Fills in missing rows and/or columns in a matrix.
}
\details{
No warning is provided if x contains row or column names not in row.names
and col.names; they are automaticaly removed.
}
