% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FillInVector}
\alias{FillInVector}
\title{\code{FillInVector}}
\usage{
FillInVector(x, element.names, value = 0)
}
\arguments{
\item{x}{The vector}

\item{element.names}{The required names of the elements of the vector.}

\item{value}{The value to fill in for cells that are not in the origial matrix.}
}
\description{
Fills in missing rows and/or columns in a matrix.
}
