% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{Growth}
\alias{Growth}
\title{Growth}
\usage{
Growth(x, start, end, by, remove.last = TRUE)
}
\arguments{
\item{x}{A vector showing some metric with scale properties}

\item{start}{The data at which the analysis should start.}

\item{end}{The data at which the analysis should end.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{remove.last}{Remove the final period (as usually is incomplete).}
}
\value{
A \code{\link{list}} containing the following elements:
\item{revenue}{The revenue by \code{period}}.
\item{growth}{The revenue growth, as a proportion, by \code{period}}.
}
\description{
Sales and sales growth.
}
