% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{Heatmap}
\alias{Heatmap}
\title{\code{Heatmap}}
\usage{
Heatmap(x, title)
}
\arguments{
\item{x}{A matrix.}

\item{title}{The title of the chart.}
}
\value{
A plotly plot.
}
\description{
A heatmap of a matrix.
}
