% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{Index}
\alias{Index}
\title{\code{Index}}
\usage{
Index(x, by.row = TRUE, i = 1, STATS = NULL, remove = "",
  remove.diag = FALSE)
}
\arguments{
\item{x}{The data to be indexed.}

\item{by.row}{Index y rows.}

\item{i}{The row or column to be used as the base for the index.}

\item{STATS}{The row or column number to index by, or, a vector containing the data to be swept out.}

\item{remove}{Where the options are "", which does nothing, and "lower left",
"lower right", "upper left", and "upper right", and they are set to NA.}

\item{remove.diag}{If TRUE, and \code{remove} is not set to "", removes the diagonal as well.}
}
\description{
Indexes the data on a table by its rows or columns.
}

