% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{IndexDiagonal}
\alias{IndexDiagonal}
\title{\code{IndexDiagonal}}
\usage{
IndexDiagonal(x, off = FALSE, by.row = TRUE)
}
\arguments{
\item{x}{The data to be indexed.}

\item{off}{TRUE if indexing by the off diagonal.}

\item{by.row}{Indexes by row (versus by column).
those to the right of them, to NA.}
}
\description{
Indexes relative to the diagonal.
}
