% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialchurn.R
\name{InitialChurn}
\alias{InitialChurn}
\title{\code{InitialChurn}}
\usage{
InitialChurn(churn.cohort, remove.last = TRUE, volume = FALSE, by, ...)
}
\arguments{
\item{churn.cohort}{A \code{mattrix} showing churn by cohort.}

\item{remove.last}{Remove the final period (as usually is incomplete).}

\item{volume}{Weights the results by volume. Does nothing in this case.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{...}{Other arguments.}
}
\value{
A \code{\link{matrix}}
}
\description{
Computes the churn rate in the initial time period
}
