% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialrevenuepersubscriber.R
\name{InitialRevenuePerSubscriber}
\alias{InitialRevenuePerSubscriber}
\title{\code{InitialRevenuePerSubscriber}}
\usage{
InitialRevenuePerSubscriber(data, remove.last = FALSE, volume = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{remove.last}{Remove the final period (as usually is incomplete).}

\item{volume}{Weights the results by volume. Does nothing in this case.}
}
\value{
A \code{\link{matrix}}
}
\description{
Computes mean revenue for subscripbers in their first time period.
}
