% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetime.R
\name{Lifetime}
\alias{Lifetime}
\title{\code{Lifetime}}
\usage{
Lifetime(data, end = attr(data, "end"))
}
\arguments{
\item{data}{A \code{RevenueData} object.}

\item{end}{The date at which the last observation in the data as}
}
\value{
A \code{list} containing the following elements:
\item{total}{The total value of transactions, by \code{subscriber.from.period} and \code{period.counter}.}
\item{subscribers}{The number of subscribers by cohort.}
\item{revenue.per.subscriber}{The revenue per subsciber.}
\item{mean}{The average value of transactions, where the base is the number of subscribers
in the \code{subscriber.from.period}, by \code{subscriber.from.period} and \code{period.counter}.}
\item{cumulative}{The cumulative means.}
\item{index}{The \code{cumulative} means divided by the mean from the first period.}
}
\description{
Calculations of the lifetime revenue and length of a customers' relationships.
}
