% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetime.R
\name{LifetimeValue}
\alias{LifetimeValue}
\title{\code{LifetimeValue}}
\usage{
LifetimeValue(data, remove.last = TRUE)
}
\arguments{
\item{data}{A \code{RevenueData} object.}

\item{remove.last}{Remove the final period (as usually is incomplete).}
}
\value{
A \code{list} containing the following elements:
\item{total}{The total value of transactions, by \code{start.period} and \code{period.counter}.}
\item{mean}{The average value of transactions, where the base is the number of subscribers
in the \code{start.period}, by \code{start.period} and \code{period.counter}.}
\item{cumulative}{The cumulative means.}
\item{index}{The \code{cumulative} means divided by the mean from the first period.}
}
\description{
The value that subscribers have historically provided the firm. This is sometimes
referred to as "lifetime value". \link{\code{LifetimeValue}} estimates the value of a subscriber
based on both historical value and likely future value.
}

