% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanrecurringrevenue.R
\name{MeanRecurringRevenue}
\alias{MeanRecurringRevenue}
\title{\code{MeanRecurringRevenueByCohort}}
\usage{
MeanRecurringRevenue(data, days.to.count, by, ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{days.to.count}{The number of days after the initial commencing to include in the Recurring Revenue
calculation.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}
}
\value{
A matrix
}
\description{
Computes recurring revenue, by cohort.
}
