% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanrecurringrevenuebycohort.R
\name{MeanRecurringRevenueByCohort}
\alias{MeanRecurringRevenueByCohort}
\title{\code{MeanRecurringRevenueByCohort}}
\usage{
MeanRecurringRevenueByCohort(data, cohort.by = "year")
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{cohort.by}{The time period used in defining the cohorts: "year", "quarter", "month", "week", day".}

\item{period.by}{The time period to show value by}
}
\value{
A matrix
}
\description{
Computes recurring revenue, by cohort.
}
