% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanrecurringrevenuebycohort.R
\name{MeanRecurringRevenueByQuarterCohort}
\alias{MeanRecurringRevenueByQuarterCohort}
\title{\code{MeanRecurringRevenueByQuarterCohort}}
\usage{
MeanRecurringRevenueByQuarterCohort(data, volume = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{volume}{Does nothing.}
}
\value{
A matrix
}
\description{
Computes recurring revenue, by cohort.
}
