% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanrecurringrevenuebycohort.R
\name{MeanRecurringRevenueByYearCohort}
\alias{MeanRecurringRevenueByYearCohort}
\title{\code{MeanRecurringRevenueByYearCohort}}
\usage{
MeanRecurringRevenueByYearCohort(data, volume = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{volume}{Does nothing.}
}
\value{
A matrix
}
\description{
Computes recurring revenue, by cohort.
}
