% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcustomers.R
\name{NewCustomers}
\alias{NewCustomers}
\title{\code{NewCustomers}}
\usage{
NewCustomers(data, by = "quarter", ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time unit to plot. E.g., "month".}
}
\value{
A \code{\link{list}} containing the following elements:
  \item{id}{The \code{id} values of subscribers to churn.}
  \item{base}{The number of subscribers to renew or churn in the time period.}
  \item{counts}{Number of NewCustomerss by period (or in $ if \code{volume} is \code{TRUE}}.
  \item{rates}{The percentage to churn (weighted if the counts are weighted)}.
}
\description{
Computes NewCustomers, by cohort.
}
