% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PeriodNameToDate}
\alias{PeriodNameToDate}
\title{\code{PeriodNameToDate}}
\usage{
PeriodNameToDate(x, by)
}
\arguments{
\item{x}{The vector of \code{char} to convert.}

\item{by}{The time aggregation.}
}
\description{
Converts a vector of period names from a \code{\link{RevenueData}} object
into a date.
}

