% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churn.R
\name{plot.Churn}
\alias{plot.Churn}
\title{#' @export
print.Churn <- function(x, ...)
{
    printWithoutAttributes(x)
}}
\usage{
\method{plot}{Churn}(x, ...)
}
\description{
#' @export
print.Churn <- function(x, ...)
{
    printWithoutAttributes(x)
}
}
