% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanrecurringrevenue.R
\name{plot.MeanRecurringRevenue}
\alias{plot.MeanRecurringRevenue}
\title{#' @export
print.MeanRecurringRevenue <- function(x, ...)
{
    attr(x, "detail") <- NULL
    attr(x, "volume") <- NULL
    attr(out, "days.to.count") <- NULL
    class(x) <- class(x)[-1]
    print(x)
}}
\usage{
\method{plot}{MeanRecurringRevenue}(x, ...)
}
\description{
#' @export
print.MeanRecurringRevenue <- function(x, ...)
{
    attr(x, "detail") <- NULL
    attr(x, "volume") <- NULL
    attr(out, "days.to.count") <- NULL
    class(x) <- class(x)[-1]
    print(x)
}
}
