% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{plot.Waterfall}
\alias{plot.Waterfall}
\title{plot.Waterfall}
\usage{
\method{plot}{Waterfall}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Waterfall}.}

\item{...}{Additional parameters.}
}
\description{
Creates a waterfall chart showing the source(s) of change in sales for one period, relative to the previous period.
}
