% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsubgroups.R
\name{PlotSubGroups}
\alias{PlotSubGroups}
\title{\code{PlotSubGroups}}
\usage{
PlotSubGroups(data, INDICES, FUN, ...)
}
\arguments{
\item{data}{an R object, normally a data frame, possibly a matrix.}

\item{INDICES}{A character vecto indicating elements of data to be used as
variable for constructing sub-plots..}

\item{FUN}{A function to be applied to (usually data-frame) subsets of data. 
This needs to return an objet for which a plot method exist which returns
a plotly plot.}

\item{...}{Further arguments to FUN.}
}
\value{
A plotly plot#
}
\description{
Creates a small multiple plot by sub-groups
}
