% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricesensitivitymeter.R
\name{PriceSensitivityMeter}
\alias{PriceSensitivityMeter}
\title{Plots the Van Westendorp Price Sensitivity Meter}
\usage{
PriceSensitivityMeter(
  x,
  check.prices.ordered = FALSE,
  weights = NULL,
  likelihood.scale = c(0, 0.1, 0.3, 0.5, 0.7),
  output = c("Attitude of respondents", "Likelihood to buy", "Revenue",
    "Likelihood to buy and Revenue")[1],
  resolution = NULL,
  colors = c("#FF0000", "#FF0000", "#008000", "#008000"),
  line.type = c("dot", "solid", "solid", "dot"),
  line.thickness = c(1, 2, 2, 1),
  currency = "$",
  global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title.font.size = 16,
  subtitle.font.size = 12,
  footer.font.size = 8,
  legend.font.size = 10,
  hovertext.font.size = 11,
  y.title.font.size = 12,
  y2.title.font.size = 12,
  x.title.font.size = 12,
  y2.tick.font.size = 10,
  y.tick.font.size = 10,
  x.tick.font.size = 10,
  data.label.font.size = 10,
  hovertext.font.family = global.font.family,
  y.title.font.family = global.font.family,
  y.title.font.color = global.font.color,
  y.tick.font.family = global.font.family,
  y.tick.font.color = global.font.color,
  y2.title.font.family = global.font.family,
  y2.title.font.color = global.font.color,
  y2.tick.font.family = global.font.family,
  y2.tick.font.color = global.font.color,
  y2.bounds.minimum = NULL,
  y2.bounds.maximum = NULL,
  x.title = "Price",
  y2.title = "Revenue",
  y2.tick.prefix = currency,
  y2.tick.suffix = "",
  x.tick.prefix = currency,
  x.hovertext.format = ".2f",
  y.title = "",
  y.tick.format = "",
  y2.tick.format = ".2f",
  intersection.show = TRUE,
  intersection.arrow.color = global.font.color,
  intersection.arrow.size = 0,
  intersection.arrow.width = 0.7,
  intersection.arrow.length = 10,
  intersection.arrow.standoff = 0,
  intersection.label.font.family = global.font.family,
  intersection.label.font.color = global.font.color,
  intersection.label.font.size = 10,
  intersection.label.decimals = 2,
  intersection.label.wrap = TRUE,
  intersection.label.wrap.nchar = 21,
  font.units = "px",
  ...
)
}
\arguments{
\item{x}{Input table containing survey responses in 4 columns
At what price would you consider this product/brand to be
1) Very cheap, 2) Cheap, 3) Expensive, 4) Very expensive.}

\item{check.prices.ordered}{Check that prices supplied in the first 4 columns of \code{x}
are supplied in increasing order. For backwards compatibility this is off by default.}

\item{weights}{A numeric vector with length equal to the number of rows in \code{x}.
They are applied whem computing the proportions of respondents for each question}

\item{likelihood.scale}{Used in NSM calculation to convert likelihood scale to probabiliy.
Default scale assumes a 7 point scale.}

\item{output}{Type of data to show in the chart. One of "Attitude of respondents",
"Likelihood to buy", "Revenue", or "Likelihood to buy and Revenue".}

\item{resolution}{Numeric; controls the intervals (in terms of price) between which
"Proportion of respondents" is computed. For example, set to \code{0.1}, to
evaluate proportions every 10 cents. By default, we use observed values.}

\item{colors}{Character; a vector containing one or more colors specified as hex codes.}

\item{line.type}{Character; one of 'solid', 'dot', 'dashed'.}

\item{line.thickness}{Thickness, in pixels, of the series line}

\item{currency}{Character; Currency symbol to prepend to the intersection labels. These
will also be used to set the default prefix to the x tick labels and hovertext.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{title.font.size}{Integer; Title font size; default = 10.}

\item{subtitle.font.size}{Integer; subtitle font size}

\item{footer.font.size}{Integer; footer font size}

\item{legend.font.size}{Integer; Legend font size.}

\item{hovertext.font.size}{Font size of hover text.}

\item{y.title.font.size}{Integer; y-axis title font size}

\item{y2.title.font.size}{Font size of title for secondary axis. Only used wihen output is "Likelihood to bug and Revenue".}

\item{x.title.font.size}{Integer; x-axis title font size}

\item{y2.tick.font.size}{Font size of tick for secondary axis. Only used wihen output is "Likelihood to bug and Revenue".}

\item{y.tick.font.size}{Integer; y-axis tick label font size}

\item{x.tick.font.size}{Integer; x-axis tick label font size}

\item{data.label.font.size}{Integer; Font size for data label.px.}

\item{hovertext.font.family}{Font family of hover text.}

\item{y.title.font.family}{Character; y-axis title font family}

\item{y.title.font.color}{y-axis title font color as a named color in
character format (e.g. "black") or a hex code.}

\item{y.tick.font.family}{Character; y-axis tick label font family}

\item{y.tick.font.color}{y-axis tick label font color as a named color
in character format (e.g. "black") or an a hex code.}

\item{y2.title.font.family}{Font family of title for secondary axis. Only used wihen output is "Likelihood to bug and Revenue".}

\item{y2.title.font.color}{Font color of title for secondary axis. Only used wihen output is "Likelihood to bug and Revenue".}

\item{y2.tick.font.family}{Font family of tick for secondary axis. Only used wihen output is "Likelihood to bug and Revenue".}

\item{y2.tick.font.color}{Font color of tick for secondary axis. Only used wihen output is "Likelihood to bug and Revenue".}

\item{y2.bounds.minimum}{Lower bound of the range for the secondary y axis.}

\item{y2.bounds.maximum}{Upper bound of the range for the secondary y axis.}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y2.title}{Title for secondary axis. Only used wihen output is "Likelihood to bug and Revenue".}

\item{y2.tick.prefix}{Prefix for for tick labels on the secondary axis.}

\item{y2.tick.suffix}{Suffix for for tick labels on the secondary axis.}

\item{x.tick.prefix}{x-axis tick label prefix}

\item{x.hovertext.format}{A string representing a d3 formatting code.
See https://github.com/mbostock/d3/wiki/Formatting#numbers}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.tick.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{y2.tick.format}{Tick format for for tick labels on the secondary axis as a d3 format string.}

\item{intersection.show}{Logical; Whether to show labels to the intersection points of the lines.}

\item{intersection.arrow.color}{Color of the arrows to the intersection points.}

\item{intersection.arrow.size}{Size of the arrows to the intersection points.}

\item{intersection.arrow.width}{Width of the arrow heads.}

\item{intersection.arrow.length}{Scaling factor controlling length of arrows.}

\item{intersection.arrow.standoff}{Distance between arrowhead and intersection point.}

\item{intersection.label.font.family}{Character; intersection.label font family}

\item{intersection.label.font.color}{intersection.label font color as a named color in
character format (e.g. "black") or an a hex code.}

\item{intersection.label.font.size}{Integer; intersection label font size}

\item{intersection.label.decimals}{Integer; number of decimals to show on intersection label.}

\item{intersection.label.wrap}{Logical; whether the intersection label text should be wrapped.}

\item{intersection.label.wrap.nchar}{Number of characters (approximately) in each
line of the intersection label when \code{intersection.label.wrap} \code{TRUE}.}

\item{font.units}{One of "px" of "pt". By default all font sizes are specified in terms of
pixels ("px"). But changing this to "pt" will mean that the font sizes will be in terms
points ("pt"), which will be consistent with font sizes in text boxes.}

\item{...}{Other charting parameters passed to \code{\link[flipStandardCharts]{Line}}.}
}
\description{
Plots the Van Westendorp Price Sensitivity Meter
}
