% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{q.invoice.lines}
\alias{q.invoice.lines}
\title{Variables from Q's invoicelines table.}
\format{data.frame}
\usage{
q.invoice.lines
}
\description{
Information about Q's subscriber history.
}
\keyword{datasets}

