% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthaccounting.R
\name{QuickRatioPlot}
\alias{QuickRatioPlot}
\title{QuickRatioPlot}
\usage{
QuickRatioPlot(x, supress.first = 1)
}
\arguments{
\item{x}{A RevenueGrowthAccounting object.}

\item{supress.first}{The periods to suppress from the beginning of the data}
}
\description{
QuickRatioPlot
}

