% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurringrevenue.R
\name{RecurringRevenue}
\alias{RecurringRevenue}
\title{\code{RecurringRevenue}}
\usage{
RecurringRevenue(data, end = Sys.time(), by = "month")
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{end}{The date on and after which which revenue is ignored.
Note that the default value, \code{Sys.time()} may not be in the same time zone as your other data, and this
can cause unexpected results.}

\item{by}{The time unit. E.g., "month".}
}
\value{
A vector showing number of subscribers over time.
}
\description{
Computes recurring revenue, by cohort.
}
\details{
Computed based on being a subscribed on the last second of the time period. 
Partial revenue is multipled out. For exmaple, if with yearly data, if a customer has an annual license for $1,000
from 1 February to 1 February of the next year, and a second license which starts on 1 January of the last year
with a prorated price of $1000/12, then the recurring revenue is $2,000.
}
