% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churnbycohort.R
\name{RecurringRevenueChurnByCohort}
\alias{RecurringRevenueChurnByCohort}
\title{\code{RecurringRevenueChurnByCohort}}
\usage{
RecurringRevenueChurnByCohort(data, remove.last = FALSE, by, ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{remove.last}{Remove the final period (as usually is incomplete).}

\item{by}{The time unit to plot. E.g., "month".}

\item{...}{Other arguments.}
}
\value{
A \code{\link{matrix}}
}
\description{
Computes the recurring revenue churn by time period and time-based cohort
}
