% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeIncompleteSubscriptions}
\alias{removeIncompleteSubscriptions}
\title{\code{removeIncompleteSubscriptions}}
\usage{
removeIncompleteSubscriptions(data)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}
}
\description{
Removes from the data any subscriptions that have yet to be completed.
}
