% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeLast}
\alias{removeLast}
\title{\code{removeLast}}
\usage{
removeLast(data)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}
}
\description{
Removes the final data period from a \code{RevenueData} object.
}
