% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{Retention}
\alias{Retention}
\title{\code{Retention}}
\usage{
Retention(data, by, ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{...}{Other arguments.}
}
\value{
A \code{\link{list}} containing the following elements:
  \item{counts}{The number of subscribers per \code{period} by \code{subscriber.from.period}}
  \item{index}{The percentage (proportion * 100) of subscribers to remain subscribers}
  \item{retention}{Retention per \code{period} by \code{subscriber.from.period}}
  \item{retention.volume}{Same as \code{retention}, except weighted by
   the volumne (revenue) in the previous subscription renewal.}
  \item{average.retention}{Average subscriber retention.}
  \item{churn}{subscriber churn (1 - average.retention).}
  \item{average.life.span}{Estimated average subscriber lifespan (1 / churn) subscriber retention.}
  \item{average.volume}{Retention, weighted by subscriber value in the preceeding period.}
  \item{churn.volume}{Churn, weighted by subscriber value in the preceeding period.}
}
\description{
Computes retention, by cohort.
}
\details{
Where subscribers suspends their purchasing for a period, but purchases again later, the subscriber
is asumed to have been retained during the period where the account was suspended. Where
a subscriber for some reason had a subscription length that was different to the specified.
}
