% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenuebycohort.R
\name{RevenueByCohort}
\alias{RevenueByCohort}
\title{\code{RevenueByCohort}}
\usage{
RevenueByCohort(data, volume = TRUE, remove.last = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{volume}{Weights the results by volume. FALSE doesn't do anything.}

\item{remove.last}{Remove the final period (as usually is incomplete).}
}
\value{
A table.
}
\description{
Revenue by year and start year, as a stacked revenue chart.
}
