% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenuedata.R
\name{RevenueDataPreparation}
\alias{RevenueDataPreparation}
\title{RevenueDataPreparation}
\usage{
RevenueDataPreparation(
  value,
  from,
  to,
  start,
  end,
  id,
  subscription.length,
  subset,
  trim.id
)
}
\arguments{
\item{value}{A vector of containing the revenue per transaction..}

\item{from}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription commences.}

\item{to}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription ends}

\item{start}{The date at which the analysis outputs should
commence. By default, the earliest date recorded in
\code{from}.}

\item{end}{The date at which the analysis ends, which is used to
determine churn.  By default, the most recent date recorded in
\code{from}.}

\item{id}{A vector of \code{character}, unique identifier for
subscribers that made the transactions (e.g., email addresses,
names, subscriber keys).}

\item{subscription.length}{The time unit that describes the
subscription length: \code{year} to view the data by year,
\code{quarter}, and \code{month}. This is assumed to be the
billing period when determining if subscribers have churned or
not. Defaults to \code{"year"}.}

\item{subset}{An optional vector specifying a subset of
observations to be used in the calculations}

\item{trim.id}{The maximum length of the strings to be used showing
ID names (used to avoid situations where string names are so
long as to make reading of tables impossible. Defaults to 50.}
}
\value{
A \code{\link{data.frame}} that satisfies the input requirements for \code{value}
in \code{\link{RevenueData}}.
}
\description{
Cleans and tidies data for use in growth accounting
    computations for a startup. Turns all dates with 29th of Feb into the 28th.
}
