% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenuemetric.R
\name{RevenueMetric}
\alias{RevenueMetric}
\title{\code{RevenueMetric}}
\usage{
RevenueMetric(
  FUN = "Acquisition",
  output = c("Plot", "Table", "List")[1],
  value,
  from,
  to,
  start = min(from),
  end = max(from),
  id,
  subscription.length = "year",
  subset = rep(TRUE, length(id)),
  profiling = NULL,
  trim.id = 50,
  ...
)
}
\arguments{
\item{FUN}{A function that calculates a metric}

\item{output}{Whether to output as a Plot, Table, or List.}

\item{value}{A vector of containing the revenue per transaction.}

\item{from}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription commences.}

\item{to}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription ends}

\item{start}{The date at which the analysis outputs should
commence. By default, the earliest date recorded in
\code{from}.}

\item{end}{The date at which the analysis ends, which is used to
determine churn.  By default, the most recent date recorded in
\code{from}.}

\item{id}{A vector of \code{character}, unique identifier for
subscribers that made the transactions (e.g., email addresses,
names, subscriber keys).}

\item{subscription.length}{The time unit that describes the
subscription length: \code{year} to view the data by year,
\code{quarter}, and \code{month}. This is assumed to be the
billing period when determining if subscribers have churned or
not.}

\item{subset}{An optional vector specifying a subset of
observations to be used in the calculations}

\item{profiling}{Separate analyses are conducted among each unique combination of these variables.}

\item{trim.id}{The maximum length of the strings to be used showing
ID names (used to avoid situations where string names are so
long as to make reading of tables impossible.}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A plotly plot#?
}
\description{
Creates a small multiple plot by sub-groups
}
