% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenuepersubscriberbycohortbytime.R
\name{RevenuePerSubscriberByCohortByTime}
\alias{RevenuePerSubscriberByCohortByTime}
\title{\code{RevenuePerSubscriberByCohortByTime}}
\usage{
RevenuePerSubscriberByCohortByTime(data, remove.last = FALSE, volume = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{remove.last}{Remove the final period (as usually is incomplete).}

\item{volume}{Weights the results by volume. Does nothing in this case.}
}
\value{
A \code{\link{matrix}}
}
\description{
Computes mean revenue by time period and time since the subscriber joined.
}
