% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthaccounting.R
\name{RevenuePerSubscriberOverTime}
\alias{RevenuePerSubscriberOverTime}
\title{#@name data-name
#@docType qApril2016
#@author Numbers International \email{tim.bock@numbers.net.au}
#@keywords data
\code{AverageRevenuePerSubscriberOverTime}}
\usage{
RevenuePerSubscriberOverTime(value, date, id, by = "year", end = Sys.time(),
  FUN = mean, trim.id = 20, tolerance = 0.1)
}
\arguments{
\item{value}{A vector of containing the revenue per transaction.}

\item{id}{A vector of \code{character}, unique identifier for Subscribers that made the transactions (e.g., email addresses, names, Subscriber keys).}

\item{by}{\code{year} to view the data by year, \code{quarter}, and \code{month}.}

\item{tolerance}{The tolerance used in determining if a Subscriber has ceased to be a Subscriber. E.g., if yearly date, a value of .1 means
that a Subscriber will be assumed to be still a Subscriber if they have purcahsed within 1 year + 10% of a year.
#' @export}

\item{data}{A vector of class \code{POSIXct}/\code{POSIXt}, recording the date/time of each transaction.}
}
\description{
Computes the  statistics for use in growth accounting of a startup.
}

