% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startupmetric.R
\name{StartupMetric}
\alias{StartupMetric}
\title{\code{StartupMetric}}
\usage{
StartupMetric(
  FUN = "Acquisition",
  as.plot = TRUE,
  value,
  from,
  to,
  start = min(from),
  end = max(from),
  id,
  subscription.length = "year",
  subset = rep(TRUE, length(id)),
  profiling = NULL,
  trim.id = 50
)
}
\arguments{
\item{FUN}{A function that calculates a metric}

\item{as.plot}{If \code{TRUE}, outputs the metric as a plot}

\item{value}{A vector of containing the revenue per transaction.}

\item{from}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription commences.}

\item{to}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription ends}

\item{start}{The date at which the analysis outputs should
commence. By default, the earliest date recorded in
\code{from}.}

\item{end}{The date at which the analysis ends, which is used to
determine churn.  By default, the most recent date recorded in
\code{from}.}

\item{id}{A vector of \code{character}, unique identifier for
subscribers that made the transactions (e.g., email addresses,
names, subscriber keys).}

\item{subscription.length}{The time unit that describes the
subscription length: \code{year} to view the data by year,
\code{quarter}, and \code{month}. This is assumed to be the
billing period when determining if subscribers have churned or
not.}

\item{subset}{An optional vector specifying a subset of
observations to be used in the calculations}

\item{profiling}{Separate analyses are conducted among each unique combination of these variables.}

\item{trim.id}{The maximum length of the strings to be used showing
ID names (used to avoid situations where string names are so
long as to make reading of tables impossible.}
}
\value{
A plotly plot#?
}
\description{
Creates a small multiple plot by sub-groups
}
