% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscribers.R
\name{Subscribers}
\alias{Subscribers}
\title{\code{Subscribers}}
\usage{
Subscribers(data, by = "month")
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time unit. E.g., "month".}
}
\value{
A vector showing number of subscribers over time.
}
\description{
Computes acquisition, by cohort.
}
\details{
Computed based on being a subscribed on the last second of the time period.
}

