% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{Table}
\alias{Table}
\title{\code{Table}}
\usage{
Table(formula, data, FUN)
}
\arguments{
\item{formula}{A \code{formula} where the dependent variable is the variable to be aggregated over.}

\item{data}{A \code{data.frame}.}

\item{FUN}{the function to be applied: see \code{apply} for details.}
}
\description{
Creates a crosstab by aggregating numeric data over factors.
}

