% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{Trajectories}
\alias{Trajectories}
\title{\code{Trajectories}}
\usage{
Trajectories(value, from, to, id)
}
\arguments{
\item{value}{A vector of containing the quantity of interest.}

\item{from}{A vector of class \code{POSIXct} or \code{POSIXt}, recording the date
and time each subscription commences.}

\item{to}{A vector of class \code{POSIXct} or \code{POSIXt}, recording the date
and time each subscription ends}

\item{id}{A vector of \code{character}, unique identifier for subscribers that
made the transactions (e.g., email addresses, names, subscriber keys).}
}
\value{
A \code{\link{list}}, in which each element is the cumultative value within the subscription period, and the name of the element is the date.
The subscription period is determined based on the \code{from} and \code{to} dates.
}
\description{
Computse the trajectory for each unique ID.
}
