% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Triangle}
\alias{Triangle}
\title{\code{Triangle}}
\usage{
Triangle(x, position = "lower right", diag = FALSE)
}
\arguments{
\item{x}{A matrix.}

\item{position}{One of \code{"lower left"}, \code{"lower right"},
\code{"upper left"}, or \code{"upper right"}.}

\item{diag}{Logical. If \code{TRUE}, the diagonal is included in the triangle.}
}
\description{
Lower and Upper Triangle  (left and right) part of a matrix.
}
