% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthaccounting.R
\name{Waterfall}
\alias{Waterfall}
\title{Waterfall}
\usage{
Waterfall(x, period = NULL)
}
\arguments{
\item{x}{A RevenueGrowthAccounting object.}

\item{period}{A \code{character} indicating the period to plot (relative to the previous period). If NULL,
the total, aggregated across all periods, is shown.}
}
\description{
Creates a waterfall table showing the source(s) of change in sales for one period,
relative to the previous period.
}

