% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startupmetric.R
\name{YLim}
\alias{YLim}
\title{#' @export
Data <- function(x)
{
    UseMethod("Data", x)
}}
\usage{
YLim(x)
}
\description{
#' @export
Data.default <- function(x, ...)
{
    attr(x, "detail")
}
#' @export
Detail.default <- function(x, ...)
{
    attr(x, "detail")
    sapply(x$id, paste, collapse = ", ")
}
}
