% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquisition.R
\name{Acquisition}
\alias{Acquisition}
\title{\code{Acquisition}}
\usage{
Acquisition(data, remove.last = TRUE, volume = FALSE, number.periods = 1)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process.#' @details Where subscribers suspends their purchasing for a period, but purchases again later, the subscriber}

\item{remove.last}{Remove the final period (as usually is incomplete).}

\item{volume}{Weights the results by volume.}

\item{number.periods}{The number of periods of data that are treated as being 'acquisitions'. By default, this is 1, 
which means that only sales in the initial period are counted as acquisitions. If set to 12, for example, it would
means that the first 12 periods are treated as acquisitions (e.g., the first year, if the data has been setup as monthly).
is assumed to have been retained during the period where the account was suspended.}
}
\value{
A \code{\link{list}} containing the following elements:
  \item{id}{The \code{id} values of subscribers to churn.}
  \item{base}{The number of subscribers to renew or churn in the time period.}
  \item{counts}{Number of acquisitions by period (or in $ if \code{volume} is \code{TRUE}}.
  \item{rates}{The percentage to churn (weighted if the counts are weighted)}.
}
\description{
Computes acquisition, by cohort.
}
