% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquisition.R
\name{Acquisition}
\alias{Acquisition}
\title{\code{Acquisition}}
\usage{
Acquisition(data, remove.last = TRUE, volume = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{remove.last}{Remove the final period (as usually is incomplete).}

\item{volume}{Weights the results by volume.}
}
\value{
A \code{\link{list}} containing the following elements:
  \item{id}{The \code{id} values of subscribers to churn.}
  \item{base}{The number of subscribers to renew or churn in the time period.}
  \item{counts}{Number of acquisitions by period (or in $ if \code{volume} is \code{TRUE}}.
  \item{rates}{The percentage to churn (weighted if the counts are weighted)}.
}
\description{
Computes acquisition, by cohort.
}
\details{
Where subscribers suspends their purchasing for a period, but purchases again later, the subscriber
is asumed to have been retained during the period where the account was suspended.
}

