% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churnbycohort.R
\name{ChurnByCohort}
\alias{ChurnByCohort}
\title{\code{ChurnByCohort}}
\usage{
ChurnByCohort(data, remove.last = TRUE, volume = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{remove.last}{Remove the final period (as usually is incomplete).}

\item{volume}{Weights the results by volume. Does nothing in this case.}
}
\value{
A \code{\link{matrix}}
}
\description{
Computes the churn by time period and time-based cohort
}
