% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecteddotplot.R
\name{ConnectedDotPlot}
\alias{ConnectedDotPlot}
\title{\code{ConnectedDotPlot}}
\usage{
ConnectedDotPlot(name, from, to)
}
\arguments{
\item{name}{A \code{character} vector of the names of the entities to be plotted.}

\item{from}{The dates to be plotted from.}

\item{to}{The dates to be plotted to.}
}
\description{
Plots the subscription periods as a dotplot, with lines showing when subscriptions occurred
}
