% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{Growth}
\alias{Growth}
\title{Growth}
\usage{
Growth(data, remove.last = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{remove.last}{Remove the final period (as usually is incomplete).}
}
\value{
A \code{\link{list}} containing the following elements:
\item{revenue}{The revenue by \code{period}}.
\item{growth}{The revenue growth, as a proportion, by \code{period}}.
}
\description{
Sales and sales growth.
}

