% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanrecurringrevenuebycohort.R
\name{heatmap}
\alias{heatmap}
\title{\code{heatmap}}
\usage{
heatmap(x, row.title, column.title, series.hover, base, ...)
}
\arguments{
\item{x}{A \code{matrix}}

\item{row.title}{The row title; used in the hover tooltips.}

\item{column.title}{The column title; used in the hover tooltips.}

\item{series.hover}{The description of the value that appears in each cell;
used in the hover tooltips}

\item{base}{The sample size; used in the hover tooltips}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A plotly heatmap
}
\description{
Internal heatmap function.
}
