% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialchurn.R
\name{InitialRecurringRevenueChurn}
\alias{InitialRecurringRevenueChurn}
\title{InitialRecurringRevenueChurn}
\usage{
InitialRecurringRevenueChurn(data, by, ...)
}
\description{
InitialRecurringRevenueChurn
}
