% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{LayerCake}
\alias{LayerCake}
\title{\code{LayerCake}}
\usage{
LayerCake(data)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}
}
\value{
A plotly plot.
}
\description{
Plots revenue by year and start year, as a stacked revenue chart.
}

