% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanRecurringRevenue.R
\name{MeanRecurringRevenue}
\alias{MeanRecurringRevenue}
\title{\code{MeanRecurringRevenueByCohort}}
\usage{
MeanRecurringRevenue(data, days.to.count, by)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{days.to.count}{The number of days after the initial commencing to include in the Recurring Revenue
calculation.}

\item{period.by}{The time period to show value by}
}
\value{
A matrix
}
\description{
Computes recurring revenue, by cohort.
}
