% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanrecurringrevenue.R
\name{MeanRecurringRevenue2Years}
\alias{MeanRecurringRevenue2Years}
\title{\code{MeanRecurringRevenue2Years}}
\usage{
MeanRecurringRevenue2Years(data, by, ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A matrix
}
\description{
Computes annual recurring revenue by day 730 (from the date of the first license start)
}
