% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanRecurringRevenue.R
\name{MeanRecurringRevenueInitial}
\alias{MeanRecurringRevenueInitial}
\title{\code{MeanRecurringRevenueByYearCohort}}
\usage{
MeanRecurringRevenueInitial(data, by, ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time period to show value by}

\item{...}{Don't use. There to prevent errors when unnecessary parameters are passed}
}
\value{
A matrix
}
\description{
Computes annual recurring revenue committed on the first day that the licenses start.
}
