% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churn.R
\name{RecurringRevenueChurn}
\alias{RecurringRevenueChurn}
\title{\code{RecurringRevenueChurn}}
\usage{
RecurringRevenueChurn(data, by = "quarter", ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A named vector showing churn.
}
\description{
\code{RecurringRevenueChurn}
}
