% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{Retention}
\alias{Retention}
\title{\code{Retention}}
\usage{
Retention(data, remove.last = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{remove.last}{Remove the final period (as usually is incomplete).}
}
\value{
A \code{\link{list}} containing the following elements:
  \item{counts}{The number of subscribers per \code{period} by \code{start.period}}
  \item{index}{The percentage (proportion * 100) of subscribers to remain subscribers}
  \item{retention}{Retention by per \code{period} by \code{start.period}}
  \item{average}{Average subscriber retention.}
  \item{churn}{subscriber churn (1 - retention).}
  \item{average.life.span}{Estimated average subscriber lifespan (1 / churn) subscriber retention.}
  \item{average.volume}{Retention, weighted by subscriber value in the preceeding year.}
  \item{churn.volume}{Churn, weighted by subscriber value in the preceeding year.}
}
\description{
Computes retention, by cohort.
}
\details{
Where subscribers suspends their purchasing for a period, but purchases again later, the subscriber
is asumed to have been retained during the period where the account was suspended.
}

