% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue.R
\name{Revenue}
\alias{Revenue}
\title{\code{Revenue}}
\usage{
Revenue(data, end = Sys.time(), by, ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{end}{The time past which revenue is ignored (where its \code{from} data is greater than this value.
Note that the default value, \code{Sys.time()} may not be in the same time zone as your other data, and this
can cause unexpected results.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A vector showing the revenue at specific points in time.
}
\description{
Computes total amount of license revenue at points in time.
}
\details{
Computed based on being a subscribed on the last second of the time period.
}
