% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenuedata.R
\name{RevenueData}
\alias{RevenueData}
\title{\code{RevenueData}}
\usage{
RevenueData(value, from, to, begin = min(from), end = max(from), id,
  by = "year", subset = rep(TRUE, length(id)), profiling = NULL,
  trim.id = 50)
}
\arguments{
\item{value}{A vector of containing the revenue per transaction.}

\item{from}{A vector of class \code{POSIXct} or \code{POSIXt}, recording the date
and time each subscription commences.}

\item{to}{A vector of class \code{POSIXct} or \code{POSIXt}, recording the date
and time each subscription ends}

\item{begin}{The date at which the analysis outputs should commence. By default,
the earliest date recorded in \code{from}.}

\item{end}{The date at which the analysis ends, which is used to determine churn.
By default, the most recent  date recorded in \code{from}.}

\item{id}{A vector of \code{character}, unique identifier for subscribers that
made the transactions (e.g., email addresses, names, subscriber keys).}

\item{by}{\code{year} to view the data by year, \code{quarter}, and \code{month}. This is assumed to be the billing period
when determining if subscribers have churned or not.}

\item{subset}{An optional vector specifying a subset of observations to be used in the calculations}

\item{profiling}{A \code{data.frame} containing data, unique by \code{id}, to be included in the final \code{data.frame}.
Either it must contain the unique identifiers in a variable called \code{id}, or, the  \code{rownames} must match
the values of \code{id}.}

\item{trim.id}{The maximum length of the strings to be used showing ID names (used to avoid situations where
string names are so long as to make reading of tables impossible.}
}
\value{
A \code{\link{data.frame}} containing the following variables, along with any other variables in the \code{data}:
  \code{id}{The unique identifier.}
  \code{value}{The value of the transaction.}
  \code{from}{The commencement date of the subscription.}
  \code{to}{The end-date of the subscription.}
  \code{start} The date of first subscription.
  \code{last.from} The \code{from} date of the most recent subscription.
  \code{last.from.period} The \code{period} of the \code{last.from} date.
  \code{end} The final date to be used in any analysis.
  \code{start.period} The \code{period} of the beginning of the first subscription.
  \code{last.period} The \code{period} of the end of the most recent subscription.
  \code{churned} A \code{logical} indicating if the subscriber had ever ceased subscribing.
  \code{churned} A \code{logical} indicating if the subscriber had ceased subscribing in that period.
  \code{tenure.interval} A \code{interval} of \code{start} to \code{end}.
  \code{tenure} The number of whole periods from the begining of the first subscription
  to the end of the most recent.
  \code{period} The \code{period} of the subscription.
  \code{period.counter} The number of the period, where 0 indicates the initial period.
  \code{observation} The number of the subscription, starting from 1.
}
\description{
Computes the  statistics for use in growth accounting of a startup.
}

