% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startupmetric.R
\name{StartupMetric}
\alias{StartupMetric}
\title{\code{StartupMetric}}
\usage{
StartupMetric(
  value,
  from,
  to,
  start,
  end,
  id,
  subscription.length,
  subset,
  profiling = NULL,
  trim.id,
  FUN = "Acquisition",
  as.plot = TRUE
)
}
\arguments{
\item{value}{A vector of containing the revenue per transaction, or, a data frame, 
that contains 3 columns represeting \code{value}, \code{from},  \code{to}, and \code{id}. 
and has other parameters optionally provided as attributes.}

\item{from}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription commences.}

\item{to}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription ends}

\item{start}{The date at which the analysis outputs should
commence. By default, the earliest date recorded in
\code{from}.}

\item{end}{The date at which the analysis ends, which is used to
determine churn.  By default, the most recent date recorded in
\code{from}.}

\item{id}{A vector of \code{character}, unique identifier for
subscribers that made the transactions (e.g., email addresses,
names, subscriber keys).}

\item{subscription.length}{The time unit that describes the
subscription length: \code{year} to view the data by year,
\code{quarter}, and \code{month}. This is assumed to be the
billing period when determining if subscribers have churned or
not. Defaults to \code{"year"}.}

\item{subset}{An optional vector specifying a subset of
observations to be used in the calculations}

\item{profiling}{Separate analyses are conducted among each unique combination of these variables.}

\item{trim.id}{The maximum length of the strings to be used showing
ID names (used to avoid situations where string names are so
long as to make reading of tables impossible. Defaults to 50.}
}
\value{
A plotly plot#?
}
\description{
Creates a small multiple plot by sub-groups
}
