% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscribers.R
\name{Subscribers}
\alias{Subscribers}
\title{\code{Subscribers}}
\usage{
Subscribers(
  data,
  end = Sys.time(),
  by = "month",
  volume = FALSE,
  recurring = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{end}{The date on and after which which revenue is ignored.
Note that the default value, \code{Sys.time()} may not be in the same time zone as your other data, and this
can cause unexpected results.}

\item{by}{The time unit. E.g., "month".}

\item{volume}{The number of subscribers in terms of their value.}

\item{recurring}{If TRUE, and \code{volume} is also TRUE, computes the recurring revenue.}
}
\value{
A vector showing number of subscribers over time.
}
\description{
Computes acquisition, by cohort.
}
\details{
Computed based on being a subscribed on the last second of the time period.
}
