% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{TimeSeriesColumnChart}
\alias{TimeSeriesColumnChart}
\title{\code{TimeSeriesColumnChart}}
\usage{
TimeSeriesColumnChart(x, title = "", ytitle = "", series.name = "",
  tickformat = NULL, smooth = TRUE, ...)
}
\arguments{
\item{x}{A vector containing values, where the names indicate the dates.}

\item{title}{The title of the chart.}

\item{ytitle}{The title to show on the y-axis.}

\item{series.name}{E.g., "churn".}

\item{tickformat}{Plotlytickformat \code{tickformat}.}

\item{smooth}{Smooth the data using \code{supmsu}.}

\item{...}{Parameters to be passed to plotly.}
}
\value{
A plotly plot.
}
\description{
Plots a time series as columns, optionally smoothing the data.
}

