% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{Waterfall}
\alias{Waterfall}
\title{Waterfall}
\usage{
Waterfall(x, periods = NULL)
}
\arguments{
\item{x}{A GrowthAccounting object.}

\item{periods}{A vector of \code{character} indicating the period(s) to plot
(relative to the previous period). If NULL, the total, aggregated across all periods, is shown.}
}
\description{
Creates a waterfall table showing the source(s) of change in sales for one period,
relative to the previous period.
}
